Hotfix of mean function

+++ stupid_math.c
@@ -2,5 +2,5 @@
 
 float mean(int a, int b)
 {
-	return (a + b) / 2;
+	return (float)(a + b) / (float)2;
 }
